Moved typeValidation.js into validations folder
